move utilities to shared folder, remove duplication
