+++ databases/catdat/scripts/seed.ts
@@ -12,8 +12,7 @@ import type {
 	MorphismYaml
 } from './utils/seed.types'
 import { create_schema_hash, get_saved_schema_hash } from './utils/schema'
-import { PLURALS } from './config'
-import { STRUCTURE_TYPES, type StructureType } from '$shared/config'
+import { STRUCTURE_TYPES, type StructureType, PLURALS } from '$shared/config'
 
 const db = get_client()
