+++ src/components/StructureSelector.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 	import { goto } from '$app/navigation'
 	import { page } from '$app/state'
-	import { PLURALS } from '$lib/commons/structures'
 	import type { StructureType } from '$lib/commons/types'
-	import { STRUCTURE_TYPES } from '$shared/config'
+	import { STRUCTURE_TYPES, PLURALS } from '$shared/config'
 
 	type Props = {
 		selected_type: StructureType
