--- src/lib/commons/utils.ts
@@ -1,3 +1,0 @@
-export function is_string_array(arr: unknown): arr is string[] {
-	return Array.isArray(arr) && arr.every((el) => typeof el === 'string')
-}
