+++ src/pages/ComparisonResultPage.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import MetaData from '$components/MetaData.svelte'
-	import { get_property_url } from '$lib/commons/property.utils'
-	import { PLURALS } from '$lib/commons/structures'
+	import { get_property_url } from '$shared/property.utils'
+	import { PLURALS } from '$shared/config'
 	import type { ComparisonResult, StructureType } from '$lib/commons/types'
 	import {
 		faCheck,
