+++ src/pages/StructureListPage.svelte
@@ -2,10 +2,10 @@
 	import MetaData from '$components/MetaData.svelte'
 	import SearchFilter from '$components/SearchFilter.svelte'
 	import SuggestionForm from '$components/SuggestionForm.svelte'
-	import { normalize_text, pluralize } from '$lib/client/utils'
+	import { normalize_text, pluralize } from '$shared/utils'
 	import StructureList from '$components/StructureList.svelte'
 	import type { StructureShort, StructureType } from '$lib/commons/types'
-	import { PLURALS } from '$lib/commons/structures'
+	import { PLURALS } from '$shared/config'
 	import TagList from '$components/TagList.svelte'
 	import Fa from 'svelte-fa'
 	import { faInfoCircle } from '@fortawesome/free-solid-svg-icons'
