+++ src/routes/[type]-comparison/+page.server.ts
@@ -1,4 +1,4 @@
-import { is_structure_type } from '$lib/commons/structures'
+import { is_structure_type } from '$shared/config'
 import { fetch_structures } from '$lib/server/fetchers/structures'
 import { error } from '@sveltejs/kit'
