+++ src/routes/[type]-list/+page.server.ts
@@ -1,4 +1,4 @@
-import { is_structure_type } from '$lib/commons/structures'
+import { is_structure_type } from '$shared/config'
 import { fetch_structures_and_tags } from '$lib/server/fetchers/structures'
 import { error } from '@sveltejs/kit'
 import type { EntryGenerator } from './$types'
