+++ src/routes/[type]-properties/+page.server.ts
@@ -1,4 +1,4 @@
-import { is_structure_type } from '$lib/commons/structures'
+import { is_structure_type } from '$shared/config'
 import { fetch_grouped_properties_and_tags } from '$lib/server/fetchers/properties'
 import { error } from '@sveltejs/kit'
