+++ src/routes/[type]-properties/[tag]/+page.server.ts
@@ -1,4 +1,4 @@
-import { is_structure_type } from '$lib/commons/structures'
+import { is_structure_type } from '$shared/config'
 import { fetch_tagged_properties } from '$lib/server/fetchers/properties'
 import { error } from '@sveltejs/kit'
 import type { EntryGenerator } from './$types'
